\name{ehd}
\alias{ehd}
\non_function{}
\title{Depressive Mood Scale}
\usage{data(ehd)}
\description{
A data frame with 269 observations on the following 20 variables.
Jouvent, R et al 1988 La clinique polydimensionnelle de humeur depressive.
Nouvelle version echelle EHD  : Polydimensional rating scale of depressive mood.
Psychiatrie et Psychobiologie.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{e1}{Observed painfull sadness}
    \item{e2}{Emotional hyperexpressiveness}
    \item{e3}{Emotional instability}
    \item{e4}{Observed monotony}
    \item{e5}{Lack spontaneous expressivity}
    \item{e6}{Lack affective reactivity}
    \item{e7}{Emotional incontinence}
    \item{e8}{Affective hyperesthesia}
    \item{e9}{Observed explosive mood}
    \item{e10}{Worried gesture}
    \item{e11}{Observed anhedonia}
    \item{e12}{Felt sadness}
    \item{e13}{Situational anhedonia}
    \item{e14}{Felt affective indifference}
    \item{e15}{Hypersensibility unpleasent events}
    \item{e16}{Sensory anhedonia}
    \item{e17}{Felt affective monotony}
    \item{e18}{Felt hyperemotionalism}
    \item{e19}{Felt irritability}
    \item{e20}{Felt explosive mood}
  }
}
\source{
Jouvent, R et al 1988 La clinique polydimensionnelle de humeur depressive.
  Nouvelle version echelle EHD  : Polydimensional rating scale of depressive mood.
  Psychiatrie et Psychobiologie.}
\examples{
data(ehd)
str(ehd)
}
\keyword{datasets}

